preparation homogenization icon management
preparation for icon management homogenization
and a small improvement to poweronreset, which I use for testing.
In logisim there are more than half a dozen ways of placing icons for libraries and components. In a future PR, I'll organize this aspect of internal libraries in a more homogeneous way.
While maintaining compatibility for external libraries.
I'll also include a description in the jar file help section of the user manual.